minimum product
Launches a GUI with a list of employees from coffee_history.csv. The user can select who is getting coffee that day and updates the .csv file with the amount spent. Decision on who pays is who has paid the least total amount as stored in the csv file. Video to come